Fix speed not changing on Microsoft SharePoint
SharePoint has an event listener for the `ratechange` event that forcibly sets the rate to 1 or 0 depending on some internal variables. Fix this by temporarily removing the event listener when Skip Silence fails to change the rate and the new rate was not 0. When the new rate is set to 0 by SharePoint, it's probably for a good reason and SkipSilence should not tamper with it.

This fix should work on other sites that block changing the playback rate in a similar way but it has only been tested on SharePoint. To this end, add a new sample page, `example_page/ratechange.html`, for testing with an emulated version of the SharePoint behaviour.